Moved GetTxnOpts() to cmd-utils.go
