+++ cmd/cmd-utils.go
@@ -1,9 +1,14 @@
 package cmd
 
 import (
+	"context"
 	"errors"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/spf13/pflag"
 	"math/big"
+	"razor/core/types"
+	"razor/path"
 	"razor/utils"
 	"time"
 
@@ -79,4 +84,34 @@ func AssignAmountInWei(flagSet *pflag.FlagSet) (*big.Int, error) {
 		amountInWei = utils.GetAmountInWei(_amount)
 	}
 	return amountInWei, nil
+}
+
+func GetTxnOpts(transactionData types.TransactionOptions) *bind.TransactOpts {
+	defaultPath, err := path.GetDefaultPath()
+	utils.CheckError("Error in fetching default path: ", err)
+	privateKey := GetPrivateKey(transactionData.AccountAddress, transactionData.Password, defaultPath, UtilsStruct{
+		keystoreUtils: keystoreUtils,
+		cmdUtils:      cmdUtils,
+	})
+	if privateKey == nil {
+		utils.CheckError("Error in fetching private key: ", errors.New(transactionData.AccountAddress+" not present in razor-go"))
+	}
+	nonce, err := transactionData.Client.PendingNonceAt(context.Background(), common.HexToAddress(transactionData.AccountAddress))
+	utils.CheckError("Error in fetching pending nonce: ", err)
+
+	gasPrice := utils.GetGasPrice(transactionData.Client, transactionData.Config)
+
+	txnOpts, err := bind.NewKeyedTransactorWithChainID(privateKey, transactionData.ChainId)
+	utils.CheckError("Error in getting transactor: ", err)
+	txnOpts.Nonce = big.NewInt(int64(nonce))
+	txnOpts.GasPrice = gasPrice
+	txnOpts.Value = transactionData.EtherValue
+
+	gasLimit, err := utils.GetGasLimit(transactionData, txnOpts)
+	if err != nil {
+		log.Error("Error in getting gas limit: ", err)
+	}
+	log.Debug("Gas after increment: ", gasLimit)
+	txnOpts.GasLimit = gasLimit
+	return txnOpts
 }
